Add Phase 3 AWS Terraform stack and GHA OIDC deploy to ECR.
Locks in aws_g4dn.xlarge reduced-ladder results and the build→ECR→SSM→boot path without leaving compute enabled.

+++ serving/hf_server.py
@@ -125,11 +125,11 @@ def load_model(model_id: str, dtype: str = "float16") -> None:
     if _device == "cuda":
         _model = AutoModelForCausalLM.from_pretrained(
             model_id,
-            dtype=torch_dtype,
+            torch_dtype=torch_dtype,
             device_map={"": 0},
         )
     else:
-        _model = AutoModelForCausalLM.from_pretrained(model_id, dtype=torch_dtype)
+        _model = AutoModelForCausalLM.from_pretrained(model_id, torch_dtype=torch_dtype)
         _model = _model.to(_device)
     _model.eval()
     print("[hf] model ready", flush=True)
